uglify the emojis by using static import
