+++ gui/src/i18n/config.tsx
@@ -13,128 +13,11 @@ import { exists, readTextFile, BaseDirectory } from '@tauri-apps/plugin-fs';
 import { error } from '@/utils/logging';
 import { invoke } from '@tauri-apps/api/core';
 import { isTrayAvailable } from '@/utils/tauri';
+import { langs } from './names';
 
 export const defaultNS = 'translation';
 export const DEFAULT_LOCALE = 'en';
 const OVERRIDE_FILENAME = 'override.ftl';
-
-export const langs = [
-  {
-    emoji: (await import('@twemoji/svg/1f1e6-1f1ea.svg')).default,
-    name: 'عربى',
-    key: 'ar',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1e8-1f1ff.svg')).default,
-    name: 'Čeština',
-    key: 'cs',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1e9-1f1f0.svg')).default,
-    name: 'Dansk',
-    key: 'da',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1e9-1f1ea.svg')).default,
-    name: 'Deutsch',
-    key: 'de',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1fa-1f1f8.svg')).default,
-    name: 'English',
-    key: 'en',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f30e.svg')).default,
-    name: 'Español Latinoamericano',
-    key: 'es-419',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1ea-1f1f8.svg')).default,
-    name: 'Español España',
-    key: 'es-ES',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1ea-1f1ea.svg')).default,
-    name: 'Eesti',
-    key: 'et',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1eb-1f1ee.svg')).default,
-    name: 'Suomi',
-    key: 'fi',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1eb-1f1f7.svg')).default,
-    name: 'Français',
-    key: 'fr',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1ee-1f1f9.svg')).default,
-    name: 'Italiano',
-    key: 'it',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1ef-1f1f5.svg')).default,
-    name: '日本語',
-    key: 'ja',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1f0-1f1f7.svg')).default,
-    name: '한국어',
-    key: 'ko',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1f3-1f1f4.svg')).default,
-    name: 'Norsk bokmål',
-    key: 'nb-NO',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1f3-1f1f1.svg')).default,
-    name: 'Nederlands',
-    key: 'nl',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1f5-1f1f1.svg')).default,
-    name: 'Polski',
-    key: 'pl',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1e7-1f1f7.svg')).default,
-    name: 'Português Brasileiro',
-    key: 'pt-BR',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1f7-1f1fa.svg')).default,
-    name: 'Русский',
-    key: 'ru',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1fa-1f1e6.svg')).default,
-    name: 'Українська',
-    key: 'uk',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1fb-1f1f3.svg')).default,
-    name: 'Tiếng Việt',
-    key: 'vi',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f1e8-1f1f3.svg')).default,
-    name: '简体中文',
-    key: 'zh-Hans',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f9cb.svg')).default,
-    name: '繁體中文',
-    key: 'zh-Hant',
-  },
-  {
-    emoji: (await import('@twemoji/svg/1f97a.svg')).default,
-    name: 'Engwish~ OwO',
-    key: 'en-x-owo',
-  },
-];
 
 // AppConfig path: https://docs.rs/tauri/1.2.4/tauri/api/path/fn.config_dir.html
 // We doing this only once, don't want an override check to be done on runtime,
